Add doc comments for Action
